fix: move account-actions.ts into (dashboard) route group
Server action files belong under the routing folder, not src/lib —
matches change-password/actions.ts and login/actions.ts.
